Moved some files to new subdirectories
