+++ Distroir.CustomSDKLauncher.Core/Launchers/Customizable/AppLauncher/Factories/Basic/SdkApplication.cs
@@ -1,4 +1,4 @@
-namespace Distroir.CustomSDKLauncher.Core.Launchers.Customizable.AppLauncher.Factories
+namespace Distroir.CustomSDKLauncher.Core.Launchers.Customizable.AppLauncher.Factories.Basic
 {
     public enum SdkApplication
     {
