+++ Distroir.CustomSDKLauncher.Core/Launchers/Customizable/AppLauncher/Factories/SdkAppFactory.cs
@@ -1,5 +1,6 @@
 using System.Drawing;
 using System.IO;
+using Distroir.CustomSDKLauncher.Core.Launchers.Customizable.AppLauncher.Factories.Basic;
 
 namespace Distroir.CustomSDKLauncher.Core.Launchers.Customizable.AppLauncher.Factories
 {
